fix(utils): export formatTimestamp with null/undefined guard (#77)
Move formatTimestamp from LogLine.tsx to src/lib/utils.ts:
- Add falsy guard: returns '' for undefined, null, or empty string
- Retain unparseable-value fallback: returns the original string
- Export from utils.ts so any surface can import it
- Update LogLine.tsx to import from utils.ts instead of local copy

Add test/lib/utils.test.ts with 5 tests covering the guard cases
and the happy-path HH:MM:SS.mmm format output.
