+++ packages/webview/src/components/LogLine.tsx
@@ -15,7 +15,7 @@
 
 import { type JSX } from "preact";
 import { cva, type VariantProps } from "class-variance-authority";
-import { cn } from "../lib/utils";
+import { cn, formatTimestamp } from "../lib/utils";
 
 // ── Level colours ─────────────────────────────────────────────────────────────
 
@@ -57,16 +57,6 @@ export interface LogLineProps extends VariantProps<typeof lineVariants> {
 }
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
-
-function formatTimestamp(iso: string): string {
-  const d = new Date(iso);
-  if (Number.isNaN(d.getTime())) return iso;
-  const h = String(d.getHours()).padStart(2, "0");
-  const m = String(d.getMinutes()).padStart(2, "0");
-  const s = String(d.getSeconds()).padStart(2, "0");
-  const ms = String(d.getMilliseconds()).padStart(3, "0");
-  return `${h}:${m}:${s}.${ms}`;
-}
 
 // ── Component ─────────────────────────────────────────────────────────────────
