+++ packages/webview/src/lib/utils.ts
@@ -1,12 +1,17 @@
 import { clsx, type ClassValue } from "clsx";
 import { twMerge } from "tailwind-merge";
 
-/**
- * Merges Tailwind CSS class names, deduplicating conflicting utilities.
- *
- * Uses `clsx` for conditional class assembly and `tailwind-merge` to resolve
- * conflicts (e.g. `bg-red-500 bg-blue-500` → `bg-blue-500`).
- */
 export function cn(...inputs: ClassValue[]): string {
   return twMerge(clsx(inputs));
+}
+
+export function formatTimestamp(iso: string): string {
+  if (!iso) return "";
+  const d = new Date(iso);
+  if (Number.isNaN(d.getTime())) return iso;
+  const h = String(d.getHours()).padStart(2, "0");
+  const m = String(d.getMinutes()).padStart(2, "0");
+  const s = String(d.getSeconds()).padStart(2, "0");
+  const ms = String(d.getMilliseconds()).padStart(3, "0");
+  return `${h}:${m}:${s}.${ms}`;
 }
